Move App.vue to vue3
